feat: Update TTL sweeper to accept float for sweep interval
- Changed the `sweep_interval_minutes` parameter type from `int` to `float` in both `AsyncSurrealStore` and `SurrealStore` to allow for more precise interval settings.
- Enhanced tests to validate the TTL sweeper functionality, ensuring it correctly deletes expired items based on the new interval configuration.

+++ src/langgraph_surrealdb/store/aio.py
@@ -303,7 +303,7 @@ class AsyncSurrealStore(AsyncBatchedBaseStore):
             return len(deleted)
 
     async def start_ttl_sweeper(
-        self, sweep_interval_minutes: int | None = None
+        self, sweep_interval_minutes: float | None = None
     ) -> asyncio.Task[None]:
         """Periodically delete expired store items based on TTL.
